+++ tests/Js/Main/TimeOnlyTests.fs
@@ -154,7 +154,7 @@ let tests =
             let t1 = TimeOnly (5, 0, 2, 22)
             let t2 = TimeOnly (14, 2)
 
-//            t2.ToString() |> equal "14:02"
+            // t2.ToString() |> equal "14:02"
 
             t1.ToString(CultureInfo.InvariantCulture) |> equal "05:00"
             t2.ToString(CultureInfo.InvariantCulture) |> equal "14:02"